pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tutor-tui"
version = "0.1.0"
description = "Tutor de cursos técnicos con TUI y LLM local (Ollama)"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Claude Code", email = "noreply@anthropic.com"}
]
keywords = ["education", "tui", "llm", "ollama", "tutor"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education :: Computer Aided Instruction (CAI)",
]
dependencies = [
    "textual>=0.46.0",
    "textual-plotext>=0.2.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0.1",
    "rich>=13.0.0",
    "pydantic>=2.5.0",
    "click>=8.1.0",
    "platformdirs>=4.0.0",
    "markdown>=3.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-textual-snapshot>=0.4.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
]

[project.scripts]
tutor = "tutor_tui.__main__:main"
tutor-tui = "tutor_tui.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tutor_tui"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
